+++ R/value_extraction.R
@@ -19,7 +19,7 @@
 #' @param object A \code{\link{lm}} object, or \code{\link{formula}}.
 #' @param data If \code{object} is a formula, the data to fit the formula to as
 #'   a \code{\link{data.frame}}.
-#' @param ... Passthrough arguments to \code{\link{lm}}.
+#' @param ... Additional arguments passed through to \code{\link{lm}}.
 #'
 #' @return The value of the estimate as a single number.
 #'
@@ -58,14 +58,14 @@ PRE <- function(object, data = NULL, ...) {
 #' @export
 SSE <- function(object, data = NULL, ...) {
   fit <- convert_lm(object, data, ...)
-  sum(fit$residuals ^ 2)
+  sum(fit$residuals^2)
 }
 
 #' @rdname estimate_extraction
 #' @export
 SSM <- function(object, data = NULL, ...) {
   fit <- convert_lm(object, data, ...)
-  sum((fit$fitted.values - mean(fit$model[[1]])) ^ 2)
+  sum((fit$fitted.values - mean(fit$model[[1]]))^2)
 }
 
 #' @rdname estimate_extraction