+++ tests/testthat/test-supernova.r
@@ -1,5 +1,4 @@
-context("ANOVA values and formatting")
-library(dplyr)
+context("supernova: Values, independent designs")
 
 # Notes -------------------------------------------------------------------
 
@@ -16,34 +15,43 @@ library(dplyr)
 # Helper functions --------------------------------------------------------
 
 pre <- function(ssr, sse) ssr / (ssr + sse)
-ssr <- function(model) sum((predict(model) - mean(model$model[,1])) ^ 2, na.rm = TRUE)
-sse <- function(model) sum(resid(model) ^ 2, na.rm = TRUE)
-sst <- function(model) var(model$model[,1], na.rm = TRUE) * (nrow(model$model) - 1)
+ssr <- function(model) sum((predict(model) - mean(model$model[, 1]))^2, na.rm = TRUE)
+sse <- function(model) sum(resid(model)^2, na.rm = TRUE)
+sst <- function(model) var(model$model[, 1], na.rm = TRUE) * (nrow(model$model) - 1)
 formula_to_string <- function(formula) Reduce(paste, deparse(formula))
 
 # Retrieves partial row data for multiple regression from the model cache
 get_partials <- function(model, type) {
-  ivs <- formula(model) %>% terms() %>% labels()
-  if (length(ivs) < 2) return(NULL)
+  ivs <- formula(model) %>%
+    terms() %>%
+    labels()
+  if (length(ivs) < 2) {
+    return(NULL)
+  }
 
   if (!exists("model_cache")) {
     model_cache <<- paste0("./cache/model_cache_type_", 1:3, ".Rds") %>%
-      purrr::map(readRDS)
+      lapply(readRDS)
   }
 
   cache_name <- formula_to_string(model$call)
 
-  model_cache[[type]][[cache_name]] %>%
-    select(SS = `Sum Sq`, df = Df, F = `F value`, p = `Pr(>F)`) %>%
-    mutate(term = rownames(.)) %>%
-    filter(term %in% ivs)
+  anova_tbl <- model_cache[[type]][[cache_name]] %>%
+    .[c("Sum Sq", "Df", "F value", "Pr(>F)")] %>%
+    setNames(c("SS", "df", "F", "p")) %>%
+    {
+      .$term <- rownames(.)
+      .
+    }
+
+  anova_tbl[which(anova_tbl$term %in% ivs), ]
 }
 
 # This needs to match with df.missing in cache_test_data.R
 get_data_with_missing <- function() {
   df.missing <- mtcars
-  df.missing[1,]$hp <- NA_real_
-  df.missing[2:3,]$disp <- NA_real_
+  df.missing[1, ]$hp <- NA_real_
+  df.missing[2:3, ]$disp <- NA_real_
   return(df.missing)
 }
 
@@ -65,7 +73,7 @@ expect_col_equal <- function(object, col_name, expected, ...) {
   comp <- compare(act$val, exp$val, ...)
   expect(comp$equal, sprintf(
     "%s not equal to %s.\n%s\nActual: %s\n", "Expected: %s",
-    act$lab, exp$lab, comp$message, tibble(act$val), tibble(exp$val)
+    act$lab, exp$lab, comp$message, tibble::tibble(act$val), tibble::tibble(exp$val)
   ))
   invisible(object)
 }
@@ -83,9 +91,7 @@ expect_col_equal <- function(object, col_name, expected, ...) {
 #' @return Invisibly returns the original `supernova` object for piping.
 expect_supernova <- function(object, type = 3, ...) {
   expect_model(object)
-  if (nrow(object$tbl) > 3) {
-    expect_partials(object, type, ...)
-  }
+  if (nrow(object$tbl) > 3) expect_partials(object, type, ...)
   invisible(object)
 }
 
@@ -105,17 +111,18 @@ expect_model <- function(object, ...) {
   p_value <- pf(f_stats[[1]], f_stats[[2]], f_stats[[3]], lower.tail = FALSE)
   pre <- summary(object$fit)[["r.squared"]]
 
-  exp <- tibble(
+  expected <- data.frame(
     SS = c(ssr(object$fit), sse(object$fit), sst(object$fit)),
     df = c(f_stats[[2]], f_stats[[3]], nrow(object$fit$model) - 1),
-    MS = SS / df,
     F = c(f_stats[[1]], NA_real_, NA_real_),
     PRE = c(pre, NA_real_, NA_real_),
     p = c(p_value, NA_real_, NA_real_)
-  ) %>% as.data.frame()
+  )
+  expected["MS"] <- expected["SS"] / expected["df"]
+  expected <- expected[c("SS", "df", "MS", "F", "PRE", "p")]
 
   act_model_rows <- object$tbl %>% .[c(1, nrow(.) - 1, nrow(.)), 3:8]
-  expect_table_data(act_model_rows, exp)
+  expect_table_data(act_model_rows, expected)
   invisible(object)
 }
 
@@ -132,10 +139,13 @@ expect_model <- function(object, ...) {
 #' @return Invisibly returns the original `supernova` object for piping.
 expect_partials <- function(object, type = 3, ...) {
   act_partial_rows <- object$tbl %>% .[2:(nrow(.) - 2), 3:8]
-  exp <- get_partials(object$fit, type = type) %>%
-    mutate(MS = SS / df, PRE = pre(SS, sse(object$fit))) %>%
-    select(SS, df, MS, F, PRE, p)
-  expect_table_data(act_partial_rows, exp)
+
+  expected <- get_partials(object$fit, type = type)
+  expected["MS"] <- expected["SS"] / expected["df"]
+  expected["PRE"] <- pre(expected["SS"], sse(object$fit))
+  expected <- expected[c("SS", "df", "MS", "F", "PRE", "p")]
+
+  expect_table_data(act_partial_rows, expected)
   invisible(object)
 }
 
@@ -172,11 +182,17 @@ test_that("superanova is an alias of supernova", {
 test_that("supernova object has table, fit, and models", {
   fit <- lm(mpg ~ NULL, mtcars)
   obj <- supernova(fit, type = 3)
-  obj %>% expect_is("supernova")
-  obj$tbl %>% expect_is("data.frame")
+
+  obj %>%
+    expect_is("supernova")
+
+  obj$tbl %>%
+    expect_is("data.frame")
+
   obj$fit %>%
     expect_is("lm") %>%
     expect_identical(fit)
+
   obj$models %>%
     expect_is("comparison_models") %>%
     expect_identical(suppressWarnings(generate_models(fit, 3)))
@@ -184,6 +200,7 @@ test_that("supernova object has table, fit, and models", {
 
 test_that("supernova table structure is well-formed", {
   obj <- supernova(lm(mpg ~ NULL, mtcars))$tbl %>% expect_is("data.frame")
+  expect_named(obj, c("term", "description", "SS", "df", "MS", "F", "PRE", "p"))
   expect_is(obj[["term"]], "character")
   expect_is(obj[["description"]], "character")
   expect_is(obj[["SS"]], "numeric")
@@ -195,7 +212,9 @@ test_that("supernova table structure is well-formed", {
 })
 
 test_that("magrittr can pipe lm() to supernova", {
-  expect_is(supernova(lm(mpg ~ NULL, mtcars)), "supernova")
+  lm(mpg ~ NULL, mtcars) %>%
+    supernova() %>%
+    expect_is("supernova")
 })
 
 test_that("magrittr can pipe data to lm() to supernova", {
@@ -209,44 +228,49 @@ test_that("magrittr can pipe data to lm() to supernova", {
 })
 
 
-# Regression values -------------------------------------------------------
+# ANOVA values ------------------------------------------------------------
 
 test_that("supernova calcs. (quant. ~ NULL) ANOVA correctly", {
   model <- lm(Thumb ~ NULL, Fingers)
-  exp <- anova(model) %>%
-    mutate(F = NA_real_, PRE = NA_real_, p = NA_real_) %>%
-    select(SS = `Sum Sq`, df = Df, MS = `Mean Sq`, F, PRE, p)
-  expect_table_data(supernova(model)$tbl %>% .[nrow(.), 3:8], exp)
+  expected <- anova(model)
+  expected[c("SS", "df", "MS")] <- expected[c("Sum Sq", "Df", "Mean Sq")]
+  expected[c("F", "PRE", "p")] <- NA_real_
+  expected <- expected[c("SS", "df", "MS", "F", "PRE", "p")]
+
+  supernova(model)$tbl %>%
+    .[nrow(.), 3:8] %>%
+    expect_table_data(expected)
 })
 
 models_to_test <- list(
-  `q ~ q`         = lm(Thumb ~ Weight, Fingers),
-  `q ~ c`         = lm(Thumb ~ RaceEthnic, Fingers),
-  `q ~ q + q`     = lm(Thumb ~ Weight + Height, Fingers),
-  `q ~ c + q`     = lm(Thumb ~ RaceEthnic + Weight, Fingers),
-  `q ~ c + c`     = lm(Thumb ~ RaceEthnic + Sex, Fingers),
+  `q ~ q` = lm(Thumb ~ Weight, Fingers),
+  `q ~ c` = lm(Thumb ~ RaceEthnic, Fingers),
+  `q ~ q + q` = lm(Thumb ~ Weight + Height, Fingers),
+  `q ~ c + q` = lm(Thumb ~ RaceEthnic + Weight, Fingers),
+  `q ~ c + c` = lm(Thumb ~ RaceEthnic + Sex, Fingers),
   `q ~ c + c + c` = lm(Thumb ~ RaceEthnic + Weight + Sex, Fingers),
-  `q ~ q * q`     = lm(Thumb ~ Weight * Height, Fingers),
+  `q ~ q * q` = lm(Thumb ~ Weight * Height, Fingers),
 
-  `q ~ c * q`     = lm(Thumb ~ RaceEthnic * Weight, Fingers),
-  `q ~ c * c`     = lm(Thumb ~ RaceEthnic * Sex, Fingers),
+  `q ~ c * q` = lm(Thumb ~ RaceEthnic * Weight, Fingers),
+  `q ~ c * c` = lm(Thumb ~ RaceEthnic * Sex, Fingers),
   `q ~ c + q * c` = lm(Thumb ~ RaceEthnic + Weight * Sex, Fingers),
   `q ~ c * q * c` = lm(Thumb ~ RaceEthnic * Weight * Sex, Fingers)
 )
 
-test_that("supernova can calculate sequential (type 1) SS", {
-  purrr::map(models_to_test, ~expect_supernova(supernova(.x, type = 1), 1))
+test_that("supernova can calculate different SS types sequential (type 1) SS", {
+  anova_tester <- function(model, type) expect_supernova(supernova(model, type), type)
+
+  anova_tester_type_1 <- function(model) anova_tester(model, type = 1)
+  lapply(models_to_test, anova_tester_type_1)
+
+  anova_tester_type_2 <- function(model) anova_tester(model, type = 2)
+  lapply(models_to_test, anova_tester_type_2)
+
+  anova_tester_type_3 <- function(model) anova_tester(model, type = 3)
+  lapply(models_to_test, anova_tester_type_3)
 })
 
-test_that("supernova can calculate hierarchical (type 2) SS", {
-  purrr::map(models_to_test, ~expect_supernova(supernova(.x, type = 2), 2))
-})
-
-test_that("supernova can calculate orthogonal (type 3) SS", {
-  purrr::map(models_to_test[8], ~expect_supernova(supernova(.x, type = 3), 3))
-})
-
-test_that("supernova SS types can be numeric, numeral, character", {
+test_that("supernova SS types can be specified as numeric, numeral, character", {
   model <- models_to_test[[length(models_to_test)]]
   expect_identical(supernova(model, type = 1), supernova(model, type = "I"))
   expect_identical(supernova(model, type = 1), supernova(model, type = "sequential"))
@@ -271,102 +295,45 @@ test_that("supernova uses listwise deletion for missing data", {
   # one-way
   df_total <- sum(!is.na(df.missing$hp)) - 1
   supernova(lm(mpg ~ hp, df.missing)) %>%
-    expect_supernova()  %>%
+    expect_supernova() %>%
     # explicitly test correct df
     # this needs to be checked because otherwise the total row will show
     # nrow() - 1 for df instead of looking at only complete cases
-    .$tbl %>% expect_col_equal("df", c(1, df_total - 1, df_total))
+    .$tbl %>%
+    expect_col_equal("df", c(1, df_total - 1, df_total))
 
   # two-way
-  df_total <- nrow(na.omit(select(df.missing, mpg, hp, disp))) - 1
-  supernova(lm(mpg ~ hp * disp, df.missing)) %>%
-    expect_supernova() %>%
-    .$tbl %>% expect_col_equal("df", c(3, 1, 1, 1, df_total - 3, df_total))
+  df_total <- nrow(na.omit(df.missing[c("mpg", "hp", "disp")])) - 1
+  anova <- supernova(lm(mpg ~ hp * disp, df.missing))
+  expect_supernova(anova)
+  expect_col_equal(anova$tbl, "df", c(3, 1, 1, 1, df_total - 3, df_total))
 })
 
 test_that("message is given for number of rows deleted due to missing cases", {
   df.missing <- get_data_with_missing()
 
-  expect_message(supernova(lm(mpg ~ hp, mtcars)), NA)
+  expect_message(
+    supernova(lm(mpg ~ hp, mtcars)),
+    NA
+  )
   expect_message(
     supernova(lm(mpg ~ hp, df.missing)),
-    "Note: 1 case removed due to missing value(s). Row number: 1",
-    fixed = TRUE)
+    "Note: 1 case removed due to missing value(s).",
+    fixed = TRUE
+  )
   expect_message(
     supernova(lm(mpg ~ disp, df.missing)),
-    "Note: 2 cases removed due to missing value(s). Row numbers: 2, 3",
-    fixed = TRUE)
+    "Note: 2 cases removed due to missing value(s).",
+    fixed = TRUE
+  )
   expect_message(
     supernova(lm(mpg ~ hp * disp, df.missing)),
-    "Note: 3 cases removed due to missing value(s). Row numbers: 1, 2, 3",
-    fixed = TRUE)
+    "Note: 3 cases removed due to missing value(s).",
+    fixed = TRUE
+  )
 })
 
 test_that("supernova makes correct calculations for unbalanced data", {
-  expect_supernova(supernova(lm(uptake ~ Treatment, data = CO2[1:80,])))
-  expect_supernova(supernova(lm(uptake ~ Treatment * Type, data = CO2[1:80,])))
+  expect_supernova(supernova(lm(uptake ~ Treatment, data = CO2[1:80, ])))
+  expect_supernova(supernova(lm(uptake ~ Treatment * Type, data = CO2[1:80, ])))
 })
-
-
-# Formatting --------------------------------------------------------------
-
-test_that("supernova tables object has explanatory header", {
-  model <- lm(mpg ~ hp * disp, data = mtcars)
-  printed <- capture.output(supernova(model, type = 1))
-  expect_identical(printed[1:3], c(
-    " Analysis of Variance Table (Type I SS)",
-    " Model: mpg ~ hp * disp",
-    " "
-  ))
-
-  printed <- capture.output(supernova(model, type = 2))
-  expect_identical(printed[[1]], " Analysis of Variance Table (Type II SS)")
-
-  printed <- capture.output(supernova(model, type = 3))
-  expect_identical(printed[[1]], " Analysis of Variance Table (Type III SS)")
-})
-
-test_that("columns have general ANOVA table names", {
-  model <- lm(mpg ~ hp * disp, data = mtcars)
-  printed <- capture.output(supernova(model, type = 1))
-  expect_match(printed[[4]], "^\\s+SS\\s+df\\s+MS\\s+F\\s+PRE\\s+p$")
-})
-
-test_that("null model tables are beautifully formatted", {
-  model <- lm(mpg ~ NULL, data = mtcars)
-  printed <- capture.output(supernova(model))
-  expect_identical(printed[5:9], c(
-    " ----- ----------------- -------- --- ------ --- --- ---",
-    " Model (error reduced) |      --- ---    --- --- --- ---",
-    " Error (from model)    |      --- ---    --- --- --- ---",
-    " ----- ----------------- -------- --- ------ --- --- ---",
-    " Total (empty model)   | 1126.047  31 36.324            "
-  ))
-})
-
-test_that("single predictor tables are beautifully formatted", {
-  model <- lm(mpg ~ hp, data = mtcars)
-  printed <- capture.output(supernova(model))
-  expect_identical(printed[5:9], c(
-    " ----- ----------------- -------- -- ------- ------ ------ -----",
-    " Model (error reduced) |  678.373  1 678.373 45.460 0.6024 .0000",
-    " Error (from model)    |  447.674 30  14.922                    ",
-    " ----- ----------------- -------- -- ------- ------ ------ -----",
-    " Total (empty model)   | 1126.047 31  36.324                    "
-  ))
-})
-
-test_that("multiple predictor tables are beautifully formatted", {
-  model <- lm(mpg ~ hp + disp, data = mtcars)
-  printed <- capture.output(supernova(model))
-  expect_identical(printed[5:11], c(
-    " ----- ----------------- -------- -- ------- ------ ------ -----",
-    " Model (error reduced) |  842.554  2 421.277 43.095 0.7482 .0000",
-    "    hp                 |   33.665  1  33.665  3.444 0.1061 .0737",
-    "  disp                 |  164.181  1 164.181 16.795 0.3667 .0003",
-    " Error (from model)    |  283.493 29   9.776                    ",
-    " ----- ----------------- -------- -- ------- ------ ------ -----",
-    " Total (empty model)   | 1126.047 31  36.324                    "
-  ))
-})
-